Feature: Dungoneering Mastery Skillcape Emote
Added the emote, only issue is spawns -1 npc temporarily. I'll come back to this

+++ combat/com/rs/game/npc/NPC.java
@@ -70,11 +70,11 @@ public class NPC extends Entity {
 	// npc masks
 	private transient Transformation nextTransformation;
 
-	public NPC(short id, WorldTile tile, Direction direction) {
+	public NPC(int id, WorldTile tile, Direction direction) {
 		this(id, tile, direction,  true, false);
 	}
 	
-	public NPC(short id, WorldTile tile, Direction direction, boolean canBeAttackFromOutOfArea) {
+	public NPC(int id, WorldTile tile, Direction direction, boolean canBeAttackFromOutOfArea) {
 		this(id, tile, direction, canBeAttackFromOutOfArea, false);
 	}
